+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Master
 * update `heartchambers_highres` task: add postprocessing to remove segmentations outside of heart (+10mm). This fixes the issue of sometimes wrong heart segmentations in the liver or spleen. Also use the `robust_crop` option per default now for the heart chambers task.
+* add support for dicom_seg output type
+* **BREAKING CHANGE**: The option `--output_type dicom` is now `--output_type dicom_rtstruct`
+* **BREAKING CHANGE**: If setting `--output_type dicom_seg` or `--output_type dicom_rtstruct` the output path (`-o`) is not treated as a directory anymore but as a file path. So if you did `-o dicom_rt_struct_output_dir` you now have to do `-o dicom_rt_struct_output_dir/segmentations.dcm`
 
 
 ## Release 2.11.0